[refactoring] : (src/todos/components/TodoCard.tsx) TodoListContainer에서 투두 아이템 코드 분리
